score

+++ Alakajam2018_Unity/Assets/_Scripts/Score.cs
@@ -8,9 +8,9 @@ public class Score : MonoBehaviour {
     private Transform player;
 
     [HideInInspector]
-    public int currentScore;
+    public float currentScore;
 
-    private int playerPos;
+    private float playerPos;
 
     public TMP_Text points;
 
@@ -21,7 +21,7 @@ public class Score : MonoBehaviour {
 	
 	void FixedUpdate () 
     {
-        playerPos = (int)player.transform.position.y;
+        playerPos = player.transform.position.y;
 
         //incase player falls down, keeps points
         if (playerPos >= currentScore)
